chore: replace lib ts
